pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "traccia"
version = "0.1.0"
description = "Local-first reflective skill graph compiler."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.11,<3",
  "pypdf>=5.4,<6",
  "pyyaml>=6.0.2,<7",
  "typer>=0.16,<1",
  "python-docx>=1.1,<2",
]

[project.scripts]
traccia = "traccia.cli:app"

[dependency-groups]
dev = [
  "pytest>=8.3,<9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/traccia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
